+++ app/api/posts/[id]/route.ts
@@ -5,7 +5,8 @@ import { Post } from '@/types/post'
 import { deriveTitleFromContent } from '@/lib/posts'
 import { SupabaseClient } from '@supabase/supabase-js'
 
-export const dynamic = 'force-dynamic'
+// Use ISR for GET requests - revalidate every 5 minutes
+export const revalidate = 300
 
 async function checkIsAdmin(supabase: SupabaseClient, userId: string): Promise<boolean> {
   const { data: roleData } = await supabase
@@ -40,7 +41,6 @@ export async function GET(
 
     return NextResponse.json(post)
   } catch (error) {
-    console.error('Error fetching post:', error)
     return NextResponse.json(
       { error: 'Internal server error' },
       { status: 500 }
@@ -129,7 +129,6 @@ export async function PUT(
 
     return NextResponse.json(post)
   } catch (error) {
-    console.error('Error updating post:', error)
     return NextResponse.json(
       { error: 'Internal server error' },
       { status: 500 }
@@ -178,7 +177,6 @@ export async function DELETE(
       { status: 200 }
     )
   } catch (error) {
-    console.error('Error deleting post:', error)
     return NextResponse.json(
       { error: 'Internal server error' },
       { status: 500 }
